add knobs for radio
